requirements.txt
arrow==0.14.6
Click==7.0
click-default-group==1.2.1
pathlib==1.0.1
inquirer==2.6.3
retrying==1.3.3
humanize==0.5.1
passlib==1.7.1
tabulate==0.8.3
psycopg2-binary==2.8.3
requests==2.20.1
requests-oauth==0.4.1
requests-oauth2==0.3.0
requests-oauthlib==1.2.0
lxml==4.4.1
psutil==5.6.3
